+++ Elearning/Equipment.cs
@@ -8,7 +8,7 @@ namespace Elearning
         public UInt16 Price { get; }
         public DateOnly RegistrationDate { get; }
 
-        public Equipment(string name, int price, DateOnly? registrationDate = null)
+        public Equipment(string name, UInt16 price, DateOnly? registrationDate = null)
         {
             Name = name;
             Price = price;
